gram/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==2.2
django-currentuser==0.5.3
django-taggit==2.1.0
easy-thumbnails==2.8.3
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
whitenoise==6.2.0
